backend/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
dj-database-url==3.1.2
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.14.0
gunicorn==25.1.0
idna==3.11
packaging==26.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2==2.9.11
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
